Обновляется только форма аутентификации при вводе почты и пароля
