+++ project/src/components/sign-in-page/sign-in-page.tsx
@@ -1,13 +1,6 @@
-import { AnyAction } from '@reduxjs/toolkit';
-import { ChangeEvent, useState } from 'react';
-import { setAuthAction } from '../../store/api-actions';
-import { store } from '../../types/state';
-import { MouseEvent } from 'react';
+import Login from '../login/login';
 
 function SignInPage() {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-
   return (
     <div className="page page--gray page--login">
       <header className="header">
@@ -24,49 +17,7 @@ function SignInPage() {
 
       <main className="page__main page__main--login">
         <div className="page__login-container container">
-          <section className="login">
-            <h1 className="login__title">Sign in</h1>
-            <form className="login__form form" action="#" method="post">
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">E-mail</label>
-                <input
-                  className="login__input form__input"
-                  type="email"
-                  name="email"
-                  value={email}
-                  onChange={(evt: ChangeEvent<HTMLInputElement>) =>
-                    setEmail(evt.target.value)}
-                  placeholder="Email"
-                  required
-                />
-              </div>
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">Password</label>
-                <input
-                  className="login__input form__input"
-                  type="password"
-                  name="password"
-                  value={password}
-                  onChange={(evt: ChangeEvent<HTMLInputElement>) =>
-                    setPassword(evt.target.value)}
-                  placeholder="Password"
-                  required
-                />
-              </div>
-              <button
-                className="login__submit form__submit button"
-                type="submit"
-                onClick={(evt: MouseEvent) => {
-                  evt.preventDefault();
-                  const tempEmail = email.trim();
-                  const tempPassword = password.trim();
-                  store.dispatch(setAuthAction({email: tempEmail, password: tempPassword}) as unknown as AnyAction);
-                }}
-              >
-                Sign in
-              </button>
-            </form>
-          </section>
+          <Login/>
           <section className="locations locations--login locations--current">
             <div className="locations__item">
               <a className="locations__item-link" href="/">
